Moved custom exceptions to separate directory
