+++ src/main/java/com/coherentsolutions/pot/insurance_service/exception/custom/BadRequestException.java
@@ -1,4 +1,4 @@
-package com.coherentsolutions.pot.insurance_service.exception;
+package com.coherentsolutions.pot.insurance_service.exception.custom;
 
 public class BadRequestException extends ApiException {
     public BadRequestException(String message) {
