+++ src/main/java/com/coherentsolutions/pot/insurance_service/exception/custom/ForbiddenException.java
@@ -1,4 +1,4 @@
-package com.coherentsolutions.pot.insurance_service.exception;
+package com.coherentsolutions.pot.insurance_service.exception.custom;
 
 public class ForbiddenException extends ApiException {
     public ForbiddenException(String message) {
