+++ src/main/java/com/coherentsolutions/pot/insurance_service/exception/custom/GenericApiException.java
@@ -1,4 +1,4 @@
-package com.coherentsolutions.pot.insurance_service.exception;
+package com.coherentsolutions.pot.insurance_service.exception.custom;
 
 public class GenericApiException extends RuntimeException {
     private final String code;
