+++ src/main/java/com/coherentsolutions/pot/insurance_service/exception/custom/InternalServerErrorException.java
@@ -1,4 +1,4 @@
-package com.coherentsolutions.pot.insurance_service.exception;
+package com.coherentsolutions.pot.insurance_service.exception.custom;
 
 public class InternalServerErrorException extends ApiException {
     public InternalServerErrorException(String message) {
