+++ src/main/java/com/coherentsolutions/pot/insurance_service/exception/custom/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package com.coherentsolutions.pot.insurance_service.exception;
+package com.coherentsolutions.pot.insurance_service.exception.custom;
 
 public class ResourceNotFoundException extends ApiException {
     public ResourceNotFoundException(String message) {
