+++ src/main/java/com/coherentsolutions/pot/insurance_service/exception/custom/ValidationException.java
@@ -1,4 +1,4 @@
-package com.coherentsolutions.pot.insurance_service.exception;
+package com.coherentsolutions.pot.insurance_service.exception.custom;
 
 public class ValidationException extends ApiException {
     public ValidationException(String message) {
